Import data associated with a list of sensors

This function imports data associated with a given list of sensor names.

importation(
  list_sensor,
  sensor_names = c("Burel-01", "Leclerc-02", "ParisMarche-03", "rueVignes-04",
    "ParisArcEnCiel-05", "RteVitre-06", "RueGdDomaine-07", "StDidierNord-08",
    "rueVallee-09", "StDidierSud-10", "RuePrieure-11", "RueCottage-12",
    "RueVeronniere-13", "RueDesEcoles-14"),
  sensor_ids = c(9000002156, 9000001906, 9000001618, 9000003090, 9000002453, 9000001844,
    9000001877, 9000002666, 9000002181, 9000002707, 9000003703, 9000003746, 9000003775,
    9000003736)
)

Arguments

list_sensor

A character vector specifying the names of sensors to import data for.

sensor_names

A character vector containing the name of each sensor that is displayed to the user

sensor_ids

A character vector containing the identifier name for each vector

Value

A data.frame containing the imported data.